Fix bug where order tracking sheet doesnt always download
didInit workaround for double mount in dev mode seems to not work when
the element has already been mounted once. Probably cos its module-level
so it is evaluted at parse time and set to true on first mount. So this
pattern really only works for things that mount once and exactly once in
your app (not a page that mounts and unmounts)